Adds volume labels to work with SELinux
When trying to use the tutorials on a system running SELinux, the
lack of volume labels causes an avc error and the tutorial files
are not readable resulting in a menu with no options.

This patch adds the Z option to the mount labels, telling Docker
to label the content with a private unshared label.